#229 Singularizes Age Rating Test Stubbing
